+++ x/wasm/keeper/keeper_test.go
@@ -423,7 +423,7 @@ func TestInstantiate(t *testing.T) {
 
 	gasAfter := ctx.GasMeter().GasConsumed()
 	if types.EnableGasVerification {
-		require.Equal(t, uint64(0x1c473), gasAfter-gasBefore)
+		require.Equal(t, uint64(0x1c479), gasAfter-gasBefore)
 	}
 
 	// ensure it is stored properly
@@ -961,7 +961,7 @@ func TestExecute(t *testing.T) {
 	// make sure gas is properly deducted from ctx
 	gasAfter := ctx.GasMeter().GasConsumed()
 	if types.EnableGasVerification {
-		require.Equal(t, uint64(0x1adb1), gasAfter-gasBefore)
+		require.Equal(t, uint64(0x1adb7), gasAfter-gasBefore)
 	}
 	// ensure bob now exists and got both payments released
 	bobAcct = accKeeper.GetAccount(ctx, bob)